If a method has a [Name] attribute then create the class' method property as a named function, rather than an anonymous function
This means that the function name will be repeated but it is beneficial
when working with third party libraries or tools that use the function
name for debugging or reporting purposes (see
https://github.com/ProductiveRage/Bridge.React/issues/13 for an
example). While not perfect, the presence of a [Name] attribute seems
like a good indicator that the name may be important.