Leverage JS engine object hash optimisations in the Dictionary when the keys are strings and the default comparer is used
The Bridge Dictionary<,> implementation uses a fairly sensible approach
of hashing the keys and then using those hash values as "buckets", in
which multiple key/value pairs may be stored. When a new bucket is
required then it may be added as a new property on a private "entries"
object, taking advantage of the fact that JS treats everything as a hash
table (and the fact that the JS engines are optimised to do so since
it's such a commonly-used pattern). However, if string keys are used and
the default comparer is being used (meaning that there are no string
"normalisation" operations required - a case-insensitive string comparer
would require some normalisation work, for example) then the work of
hashing the keys in JavaScipt may be skipped and the keys themselves
used directly as the hash table keys (while the JS engine may have to do
some hashing of its own, this will be done in optimised native code,
rather than in JavaScript). The longer that the keys are that are used,
the more benefit that these alternate code paths provide.