Updates integration test 07 to match output when all tests run.
For some reason the output of issue 2806 integration test may switch
between '$module2' and '$module3'. Currently outputs '$module2' if run
alone and '$module3' if all tests are run together. As all tests are run
at once during CI, we refreshed the expected output and added a note to
the source file.