Adds unit tests for additional issue 3593 aspects.
The unit tests checks against the side effect imposed by  the initial
implementation of the fix for this issue. Also, it helps identify if a
particular character of the long string is the cause of the issue, as it
iterates through all starts and ends with characters of the full string.